requirements.txt
aiogram>=3.6,<4.0
python-dotenv>=1.0.1
SQLAlchemy>=2.0
asyncpg>=0.29
alembic>=1.13
